+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- NodeJs Lambda function init generator
+
+### Changed
+- Bumped general Terraform AWS provider versions to v4
 
 ## [0.7.0] - 2022-09-04
 ### Fixed